Wikipedia/Crisis Misinformation Monitor/requirements.txt
crewai
python-dotenv
wikipedia
requests
pydantic
